Beginning competitive changes
Mostly just copied over ranked permissions code to comp to get the ball rolling.
I made it so that if setups are approved for ranked, they are also approved for competitive. I don't know if that's the intended functionality. It just made sense to me, a non-comp player. Might need to change that.
Also set it so that after finishing 30 ranked games (no abandons) the player is approved for comp and gets a message saying that. I have NOT tested this, so I think it works but I dunno.
Also, there should PROBABLY be more code put into Game.js for paying out different coin amounts, or increasing some kind of competitive stat in the DB. I didn't finish my changes in that script, I forgot what I had planned to do yesterday.